requirements.txt
click>=8.0.1
Flask>=3.0.0
gunicorn>=21.2.0
itsdangerous>=2.1.2
Jinja2>=3.1.2
MarkupSafe>=2.1.3
python-dotenv>=1.0.0
Werkzeug>=3.0.1
PyMySQL>=1.1.0
mysql-connector-python>=8.2.0